DataFile v3 (support timestamp of command).

+++ inc/DataFile3.rb
@@ -1,61 +1,71 @@
 class DataFile3
 
-	def initialize(fileName)
-		@fileName  = fileName 
-	end
+	def initialize(fileName, comment = nil)
+		@fileName  = fileName
 
-	def CheckCommands
+		unless File.exist?(fileName)
+			File.open(@fileName, 'a') { |file|
+				file.write("### file format:3.0 ###\n")
+			}
 
-		file = File.new(@fileName, "r")
-		while (line = file.gets)
-
-			regexp=Regexp.new("###command:(.*)###")
-		
+			unless comment.nil?
+				File.open(@fileName, 'a') { |file|
+					file.write("### comment:#{comment} ###\n")
+				}
+			end
 		end
-		file.close	
-	
+
 	end
 
 	def Write command, value
-		File.open(@fileName, 'a') { |file| 
-			file.write("###command:#{command}###\n")
+
+		File.open(@fileName, 'a') { |file|
+			file.write("### command:#{command} ### time:#{Time.now.strftime("%Y-m-d_%H:%M:%S")} ### timestamp: #{Time.now.to_i} ###\n")
 			file.write(value) 
 		}
-		#file.close
-	end 
+	end
+
+	def ListCommands
+
+		result = Hash.new
+
+		file = File.new(@fileName, 'r')
+		while (line = file.gets)
+			if match = /### command: (.*) ### time:(.*) ### timestamp: (\d+) ###/.match(line)
+				result[match[1]] = Array.new if result[match[1]].nil?
+				result[match[1]].push(match[3])
+			end
+		end
+
+		result
+	end
 
 	def Read command 
-		dataString = ""; 
-		regexpCommand=Regexp.new(Regexp.escape(command))
-		regexp=Regexp.new("###command")
+
+		result = Hash.new
 		insiteCommand=0
-		
-		file = File.new(@fileName, "r")
+		timestamp = nil
+
+		file = File.new(@fileName, 'r')
 		while (line = file.gets)
-#			puts line 
-			matchCommand = regexpCommand.match(line)
-			match = regexp.match(line)
-			
-			if match and matchCommand
-#				puts "found!!!"
-#				puts match
-				insiteCommand=1
-				next 
-			elsif match 
-#				puts "new command: #{line}"
-				insiteCommand=0
-			end	
-			
-			if insiteCommand == 1
-				line.gsub!("\r", '')
-				dataString << line 
+
+			if match = /### command: (.*) ### time:(.*) ### timestamp: (\d+) ###/.match(line)
+				match[1] == command ? insiteCommand=1 : insiteCommand=0
+				timestamp = match[3]
+				next
 			end
-			
+
+			if insiteCommand == 1
+				if result[timestamp].nil?
+					result[timestamp] = line.gsub("\r", '')
+				else
+					result[timestamp] += line.gsub("\r", '')
+				end
+			end
 		end
 		file.close	
 
-		dataString 		
+		result
 	end
-	
 
 end
